LayoutTests/media/track/track-webvtt-tc012-out-of-order.html

<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>            
 
            function trackLoaded()
            {
                findMediaElement();
                var expected = 
                {
                    length : 2,
                    tests:
                    [],
                };
                testCues(0, expected);
                
                endTest();
            }

        </script>
    </head>
    <body>
    <p>Tests that cues that are temporally out of order are ignored.</p>
        <video>
            <track src="captions-webvtt/tc012-out-of-order.vtt" onload="trackLoaded()">
        </video>
    </body>
</html>
