LayoutTests/media/track/track-webvtt-tc012-out-of-order-expected.txt

Tests that cues that are temporally out of order are ignored.


*** Testing text track 0
EXPECTED (cues.length == '2') OK
END OF TEST
